chore(SWUI-17): move Button component to its own folder
